generalized extract_new_files.sh

+++ egs/althingi/s5/local/data_extraction/extract_new_files.sh
@@ -2,14 +2,28 @@
 
 set -o pipefail
 
-. ./path.sh
-. ./conf/path.conf
+. ./path.sh # $data defined here
+. parse_options.sh || exit 1;
 
+if [ $# -ne 3 ]; then
+  echo "This script extract audio files and corresponding transcripts from althingi.is given the start and finish session numbers."
+  echo ""
+  echo "Usage: $0 <path-to-original-data> <output-data-dir>" >&2
+  echo "e.g.: $0 /data/local/corpus_20180612 132 148" >&2
+  exit 1;
+fi
 # $data is defined in path.conf
+corpusdir=$1; shift #/data/althingi/corpus_jun2018
+start=$1; shift #146
+stop=$1; #148
+
 datadir=$data/data_extraction/althingi_info
-corpusdir=/data/althingi/corpus_jun2018
-start=146
-stop=148
+
+tmp=$(mktemp -d)
+cleanup () {
+    rm -rf "$tmp"
+}
+trap cleanup EXIT
 
 # Extract info about the speeches from althingi.is
 srun --nodelist=terra sh -c "php local/data_extraction/scrape_althingi_info.php $datadir/thing $start $stop &> $datadir/log/scraping_althingi_info.log" &
@@ -27,63 +41,26 @@ done
 deactivate
 
 # All speeches already in training and testing sets:
-cat data/train_combined/text data/dev/text data/eval/text | cut -d" " -f1 | cut -d"-" -f2 | cut -d"_" -f1 | sort -u > ids_in_use
-ls ${corpusdir}/audio/ | cut -d"." -f1 | sort -u >> ids_in_use
-sort -u ids_in_use > tmp && mv tmp ids_in_use
+cat $root_intermediate/all*/text | cut -d" " -f1 | cut -d"-" -f2 | sort -u > $tmp/ids_in_use
+ls ${corpusdir}/audio/ | cut -d"." -f1 | sort -u >> $tmp/ids_in_use
 
 # Extract the ids of speeches that we don't already have in our data
-for session in $(seq $start $stop); do
-    if [ $session -lt 146 ]; then
-        comm -13 <(sort -u ids_in_use) <(cut -f1 ${datadir}/thing${session}_mp3_xml_all.txt | sort -u) > ${datadir}/new${session}.txt
-
-        while [ "$(wc -l ${datadir}/new${session}.txt | cut -d" " -f1)" -gt 3000 ]; do
-	    head -n3000 ${datadir}/new${session}.txt > ${datadir}/new${session}_head.txt
-	    sed -r -i '1,3000d' ${datadir}/new${session}.txt
-
-            # Make the regex pattern. Need to split up the lists!
-            tr "\n" "|" < ${datadir}/new${session}_head.txt | sed '$s/|$//' > ${datadir}/pattern_session${session}_head.tmp
-
-            # keep the lines that reference new files
-            sed -nr '/'$(cat ${datadir}/pattern_session${session}_head.tmp)'/p' ${datadir}/thing${session}_mp3_xml_all.txt >> ${datadir}/thing${session}_mp3_xml_new.txt
-	done
-    
-	# Make the regex pattern from the last lines.
-	tr "\n" "|" < ${datadir}/new${session}.txt | sed '$s/|$//' > ${datadir}/pattern_session${session}.tmp
-
-	# keep the lines that reference new files
-	sed -nr '/'$(cat ${datadir}/pattern_session${session}.tmp)'/p' ${datadir}/thing${session}_mp3_xml_all.txt >> ${datadir}/thing${session}_mp3_xml_new.txt
-    else
-	mv ${datadir}/thing${session}_mp3_xml_all.txt ${datadir}/thing${session}_mp3_xml_new.txt
-done
-#rm ${datadir}/*.tmp
+if [ -s $tmp/ids_in_use ]; then
+  for session in $(seq $start $stop); do
+    grep -f $tmp/ids_in_use ${datadir}/thing${session}_mp3_xml_all.txt > $tmp/used.tmp
+    comm -13 <(sort -u $tmp/used.tmp) <(sort -u ${datadir}/thing${session}_mp3_xml_all.txt) > ${datadir}/thing${session}_mp3_xml_new.txt
+  done
+else
+  mv ${datadir}/thing${session}_mp3_xml_all.txt ${datadir}/thing${session}_mp3_xml_new.txt
+fi
 
 # Download the corresponding audio and xml
 for session in $(seq $start $stop); do
     # Make sure all the files contain xml files
-    awk -F $'\t' 'NF==4{print}{}' ${datadir}/thing${session}_mp3_xml_new.txt > tmp && mv tmp ${datadir}/thing${session}_mp3_xml_new.txt
+    awk -F $'\t' 'NF==4{print}{}' ${datadir}/thing${session}_mp3_xml_new.txt > $tmp/withXML.tmp && mv $tmp/withXML.tmp ${datadir}/thing${session}_mp3_xml_new.txt
     srun --time=0-12 --nodelist=terra sh -c "php local/data_extraction/scrape_althingi_xml_mp3.php ${datadir}/thing${session}_mp3_xml_new.txt /data/althingi/corpus_jun2018/audio/ /data/althingi/corpus_jun2018/text_endanlegt/ &> ${datadir}/log/extraction_thing${session}.log" &
 done
 
-# # Since I got almost 12000 errors of the form: <html><body>Það hefur komið upp villa. Reyndu aftur síðar.</body></html>, error status 403, I try fetching them again
-# while [ "$(wc -l only_audio.txt | cut -d" " -f1)" -gt 3000 ]; do
-#     head -n3000 only_audio.txt > only_audio_head.txt
-#     sed -r -i '1,3000d' only_audio.txt
+# I got almost 12000 errors of the form: <html><body>Það hefur komið upp villa. Reyndu aftur síðar.</body></html>, error status 403. Fetching them again worked for a lot of them.
 
-#     # Make the regex pattern. Need to split up the lists!
-#     tr "\n" "|" < only_audio_head.txt | sed '$s/|$//' > only_audio_head_pattern.tmp
-
-#     # keep the lines that reference new files
-#     sed -nr '/'$(cat only_audio_head_pattern.tmp)'/p' <(cat thing*_mp3_xml_all.txt) >> only_audio_mp3_xml.txt
-# done
-    
-# # Make the regex pattern from the last lines.
-# tr "\n" "|" < only_audio.txt | sed '$s/|$//' > only_audio_pattern.tmp
-
-# # keep the lines that reference new files
-# sed -nr '/'$(cat only_audio_pattern.tmp)'/p' <(cat thing*_mp3_xml_all.txt) >> only_audio_mp3_xml.txt
-
-# # Extract the text again
-# srun --time=0-12 --nodelist=terra sh -c "php local/data_extraction/scrape_althingi_xml.php /home/staff/inga/kaldi/egs/althingi/s5/data/althingiUploads/only_audio_mp3_xml.txt &> data/althingiUploads/log/extract_xml_to_only_audio.log" &
-
-# Extract the text and put in one file. Each line is utterance ID and the utterance itself
-# utils/slurm.pl --time 0-06:00 $corpusdir/extract_text_endanlegt.log python3 local/extract_text.py $corpusdir/text_endanlegt $corpusdir/text_orig_endanlegt.txt &
+exit 0;
